nodice/pyproject.toml
[tool.poetry]
name = "nodice"
version = "0.1.0"
description = "Dicebot on qq"
authors = ["Jigsaw <j1g5aw@foxmail.com>"]
license = "AGPLv3"

[tool.poetry.dependencies]
python = "^3.7"
nonebot2 = "^2.0.0-alpha.10"
nb-cli = "^0.3.2"
PyYAML = "^5.4.1"
hjson = "^3.0.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
